requirements.txt
asgiref==3.3.4
Django==3.2.3
django-ckeditor==6.1.0
django-environ==0.4.5
django-js-asset==1.2.2
django-modeltranslation==0.17.2
Pillow==8.2.0
pip==21.1.3
python-decouple==3.4
pytz==2021.1
setuptools==57.0.0
six==1.16.0
sqlparse==0.4.1
